libpayload: Fix pointer related casts
Fix pointer related casts since this can create a problem for 64-bit systems.

BUG=None
BRANCH=None
TEST=Compiled successfully for link, nyan using emerge-* libpayload

Reviewed-on: https://chromium-review.googlesource.com/199564
Commit-Queue: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>

+++ payloads/libpayload/drivers/usb/xhci.c
@@ -127,7 +127,7 @@ xhci_init (void *bar)
 	controller->destroy_intr_queue	= xhci_destroy_intr_queue;
 	controller->poll_intr_queue	= xhci_poll_intr_queue;
 
-	controller->reg_base = (u32)bar;
+	controller->reg_base = (uintptr_t)bar;
 	controller->instance = xzalloc(sizeof(xhci_t));
 	xhci_t *const xhci = (xhci_t *)controller->instance;
